Rename article to "Why Use a Full Node Wallet"
The phrasing "Why run a full node" does not emphasize that for security
users must _use_ the full node as a wallet. Sadly many bitcoiners are
using a web wallet or SPV wallet for their economic activity and then
running a full node on rented hardware somewhere, thinking they're
helping.

The body of Pieter Wuille makes this point obvious, and the phrase "Why
run a full node" does not appear in his text.